split common_doc
