+++ test/consistency-and-style/test_summary_text.py
@@ -9,7 +9,7 @@ import pytest
 from mathics import __file__ as mathics_initfile_path
 from mathics.core.builtin import Builtin
 from mathics.core.load_builtin import name_is_builtin_symbol
-from mathics.doc.common_doc import skip_doc
+from mathics.doc.gather import skip_doc
 
 # Get file system path name for mathics.builtin
 mathics_path = osp.dirname(mathics_initfile_path)
